server/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
pyjwt
passlib[bcrypt]
